refactor to use crontab and cli
