+++ src/planscape/planning/e2e/scenario_test.py
@@ -1,0 +1,130 @@
+import logging
+import json
+import os
+from datetime import datetime
+
+from django.conf import settings
+from django.contrib.auth.models import User
+from celery import group, chain, shared_task
+from planning.models import PlanningArea, Scenario
+from planning.tasks import async_forsys_run
+from planning.e2e.tasks import review_results
+from planning.e2e.validation import load_schema
+
+log = logging.getLogger(__name__)
+
+
+@shared_task
+def run_e2e_test():
+    log.info("Running scheduled e2e test")
+    e2etest = E2EScenarioTest()
+    e2etest.initiate_tests()
+
+
+class E2EScenarioTest:
+    fixtures_to_test = []
+    scenarios_to_test = []
+    test_user_name = "e2etest@sig-gis.com"
+    fixtures_path = str(settings.TREATMENTS_TEST_FIXTURES_PATH)
+
+    def initiate_tests(self, fixtures_path=None):
+        if self.fixtures_path:
+            log.info(f"Fixtures path is set to: {self.fixtures_path}")
+        self.load_test_definitions()
+        self.upsert_test_user()
+        self.create_areas()
+        self.run_tests()
+
+    def load_test_definitions(self):
+        """Reads the JSON file that describes our test areas, scenarios and expected results"""
+        test_defs = os.path.join(
+            settings.BASE_DIR, self.fixtures_path, "test_definitions.json"
+        )
+        try:
+            with open(test_defs, "r", encoding="UTF-8") as f:
+                self.fixtures_to_test = json.load(f)
+        except Exception as e:
+            log.error(f"Failed to read test definitions at {test_defs}- {e}")
+
+    def upsert_test_user(self):
+        """Upserts the same test user for all test records"""
+        self.test_user, _ = User.objects.update_or_create(username=self.test_user_name)
+        print(
+            f"Setting test user id: {self.test_user.id} and username: {self.test_user.username}"
+        )
+
+    def create_areas(self):
+        """Loads the test planning areas under our test user"""
+        self.test_area_ids = []
+        for tests in self.fixtures_to_test:
+            planning_area_file = tests["planning_area"]
+            scenarios = tests["scenarios"]
+            areas_file = os.path.join(
+                settings.BASE_DIR,
+                self.fixtures_path,
+                planning_area_file,
+            )
+            with open(areas_file, "r", encoding="UTF-8") as f:
+                area_data = json.load(f)
+                # Remove attributes that might exist if our planning area is from the CSV output
+                if "pk" in area_data:
+                    del area_data["pk"]
+
+                # overwrite details for new planning area to be upserted
+                area_data["user_id"] = self.test_user.id
+                area_data["name"] = f"testplan-{area_data.get('name')}"
+
+                # upsert this, using name and user as unique identifier
+                # we don't want to use a PK here, because it may interfere with new data
+                area_obj, _ = PlanningArea.objects.update_or_create(
+                    user_id=area_data["user_id"],
+                    name=area_data["name"],
+                    defaults=area_data,
+                )
+                self.test_area_ids.append(area_obj.id)
+                for s in scenarios:
+                    self.upsert_scenarios(s["treatment"], s["expected"], area_obj.id)
+
+    def upsert_scenarios(self, scenario_file, validation_file, area_id):
+        """Upserts each scenario"""
+        scenario_path = os.path.join(
+            settings.BASE_DIR, self.fixtures_path, scenario_file
+        )
+        with open(scenario_path, "r", encoding="UTF-8") as f:
+            scenario_data = json.load(f)
+            # Remove attributes that might exist if our scenario is from the CSV output
+            if "planning_area" in scenario_data:
+                del scenario_data["planning_area"]
+            if "pk" in scenario_data:
+                del scenario_data["pk"]
+            if "uuid" in scenario_data:
+                del scenario_data["uuid"]
+
+            # overwrite details for new scenario to be upserted
+            scenario_data["planning_area_id"] = area_id
+            scenario_data["name"] = f"test-scenario-{scenario_data.get('name')}"
+            scenario_data["user_id"] = self.test_user.id
+
+            # upsert this, using name and user as unique identifier
+            # we don't want to use a PK here, because it may interfere with new data
+            scenario_obj, _ = Scenario.objects.update_or_create(
+                user=self.test_user, name=scenario_data["name"], defaults=scenario_data
+            )
+
+            validation_schema = load_schema(validation_file)
+            self.scenarios_to_test.append(
+                {"id": scenario_obj.id, "schema": validation_schema}
+            )
+
+    def run_tests(self):
+        """Triggers processing for each of the scenarios in our tests"""
+        all_tasks = []
+        # here we are pushing the forsys runs to Celery for processing
+        # We then chain the JSON results and send them to a validation function
+        for s in self.scenarios_to_test:
+            task = chain(
+                async_forsys_run.si(s["id"]), review_results.si(s["id"], s["schema"])
+            )
+            all_tasks.append(task)
+        task_group = group(all_tasks)
+        task_group()
