+++ src/planscape/planscape/celery.py
@@ -1,9 +1,17 @@
 import os
 
 from celery import Celery
+from celery.schedules import crontab
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "planscape.settings")
 
 app = Celery("proj")
 app.config_from_object("django.conf:settings", namespace="CELERY")
-app.autodiscover_tasks(["planning", "planning.e2e"])
+app.autodiscover_tasks(["planning", "planning.e2e", "planning.e2e.scenario_test"])
+
+app.conf.beat_schedule = {
+    "run-periodic-task": {
+        "task": "planning.e2e.scenario_test.run_e2e_test",
+        "schedule": crontab(minute="*/1"),
+    },
+}
